Move OsuJudgementInfo to its own class, remove PositionalJudgementInfo.
