+++ osu.Desktop.VisualTests/Tests/TestCaseHitObjects.cs
@@ -13,6 +13,7 @@ using osu.Game.Modes.Osu.Objects.Drawables;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.UserInterface;
+using osu.Game.Modes.Osu.Judgements;
 using OpenTK.Graphics;
 
 namespace osu.Desktop.VisualTests.Tests
