+++ osu.Game.Modes.Osu/Objects/Drawables/HitExplosion.cs
@@ -8,6 +8,7 @@ using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Transforms;
 using osu.Game.Graphics.Sprites;
 using osu.Game.Modes.Objects.Drawables;
+using osu.Game.Modes.Osu.Judgements;
 using OpenTK;
 using OpenTK.Graphics;
