+++ osu.Game.Modes.Osu/UI/OsuHitRenderer.cs
@@ -4,6 +4,7 @@
 using osu.Game.Beatmaps;
 using osu.Game.Modes.Objects.Drawables;
 using osu.Game.Modes.Osu.Beatmaps;
+using osu.Game.Modes.Osu.Judgements;
 using osu.Game.Modes.Osu.Objects;
 using osu.Game.Modes.Osu.Objects.Drawables;
 using osu.Game.Modes.UI;
